IT test fix

+++ its/plugin/src/test/java/com/sonar/cobertura/it/CoberturaTest.java
@@ -93,7 +93,7 @@ public class CoberturaTest {
             assertThat(Integer.parseInt(measureMap.get("test_errors").getValue())).isEqualTo(0);
             assertThat(Integer.parseInt(measureMap.get("skipped_tests").getValue())).isEqualTo(0);
             assertThat(Integer.parseInt(measureMap.get("test_execution_time").getValue())).isGreaterThan(0);
-            assertThat(Integer.parseInt(measureMap.get("test_success_density").getValue())).isEqualTo(100.0);
+            assertThat(Double.parseDouble(measureMap.get("test_success_density").getValue())).isEqualTo(100.0);
         }
         LOGGER.debug("Coverage result:"+measureMap.get("coverage").getValue());
         //assertThat(Float.parseFloat(measureMap.get("coverage").getValue())).isEqualTo(57.1, Delta.delta(0.1));
